user-service/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-storages
boto3
gunicorn
dj_database_url
PyJWT
Pillow
pika